fix TMT and CIDch2

+++ ms2pip/models/TMT.h
@@ -1,2 +1,2 @@
-float score_TMT_B(unsigned int* v);
-float score_TMT_Y(unsigned int* v);
+int score_TMT_B(unsigned int* v);
+int score_TMT_Y(unsigned int* v);
